Move tests into separate modules
